Moved logo.svg Added routes for users and groups Broke out the user table into it's own view
