+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from "react-redux";
+import {BrowserRouter as Router} from "react-router-dom";
 //import "./js/index";
 import './css/index.css';
 import App from './js/App';
@@ -9,9 +10,11 @@ import configureStore from './js/storage';
 
 const store = configureStore();
 ReactDOM.render(
-    <Provider store={store}>
-        <App/>
-    </Provider>,
+    <Router>
+        <Provider store={store}>
+            <App/>
+        </Provider>
+    </Router>,
     document.getElementById('root')
 );
