+++ src/js/routes/index.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import {Redirect, Route, Switch} from "react-router-dom";
+import UserTable from "../views/user-table";
+import logo from "../../images/logo.svg";
+
+export const defaultRoutes = (
+    <Switch>
+        <Route exact path="/" component={UserTable}/>
+        <Route path={'/users'} component={UserTable}/>
+        <Route path="/groups"/>
+        <Route path={'/test'} render={() => (
+            <>
+                <img src={logo} className={"App-logo"} alt={"logo"}/>
+                <h1 className={"App-title"}>Welcome</h1>
+            </>)}/>
+    </Switch>
+);
